Remove styling from #content div
Remove styling from #content div as this is the parent container of the
project. Styling this div means that every module will have to adhere
to this styling unless changes are made in the cascade. This div will
only as an anchor to append child divs/elements to the DOM.